install deps of set-login
Former-commit-id: bdf8555a45b65250daf164be0957b477ef0c7b44
Former-commit-id: d2b7d396c5ad0b75423e8b5ef4c244704a88a95f
Former-commit-id: 60d796a8ffe32590d469dc0703560b0ee95868d5

+++ source/.chezmoiscripts/post-install/run_once_after_40-set-login.sh
@@ -23,6 +23,13 @@ else
   FILE="ubuntu-gdm-set-background"
 fi
 
+# dependencies
+if ! dpkg -l | grep -q libglib2.0-dev-bin; then
+  log 'Attemping to install dependencies'
+  if [ -z "$APT" ]; then abort "Can't install libglib2.0-dev-bin using apt. Please make sure it's installed"; fi
+  sudo "$APT" install -y libglib2.0-dev-bin
+fi
+
 DESTINATION="/usr/local/bin/$FILE"
 sudo curl -SsfL "https://raw.githubusercontent.com/$OWNER/ubuntu-gdm-set-background/main/$FILE" -o "$DESTINATION"
 sudo chmod +x "$DESTINATION"
